# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface Product {
  id: string;
  name: string;
  price: number;
  availability: string;
}

export type BusinessIndustry = typeof BusinessIndustry[keyof typeof BusinessIndustry];


export const BusinessIndustry = {
  'Farm/Agriculture': 'Farm/Agriculture',
  Real_Estate: 'Real Estate',
  'E-commerce': 'E-commerce',
  Dental: 'Dental',
  Other: 'Other',
} as const;

export type BusinessConnectedChannelsItem = typeof BusinessConnectedChannelsItem[keyof typeof BusinessConnectedChannelsItem];


export const BusinessConnectedChannelsItem = {
  WhatsApp: 'WhatsApp',
  Instagram: 'Instagram',
  Email: 'Email',
  Stripe: 'Stripe',
} as const;

export type BusinessTheme = typeof BusinessTheme[keyof typeof BusinessTheme];


export const BusinessTheme = {
  green: 'green',
  navy: 'navy',
} as const;

export interface Business {
  id: string;
  name: string;
  industry: BusinessIndustry;
  website: string;
  location: string;
  timezone: string;
  currency: string;
  description: string;
  tone: string;
  avoidKeywords: string;
  connectedChannels: BusinessConnectedChannelsItem[];
  products: Product[];
  onboardingComplete: boolean;
  theme: BusinessTheme;
}

export type BusinessInputIndustry = typeof BusinessInputIndustry[keyof typeof BusinessInputIndustry];


export const BusinessInputIndustry = {
  'Farm/Agriculture': 'Farm/Agriculture',
  Real_Estate: 'Real Estate',
  'E-commerce': 'E-commerce',
  Dental: 'Dental',
  Other: 'Other',
} as const;

export type BusinessInputConnectedChannelsItem = typeof BusinessInputConnectedChannelsItem[keyof typeof BusinessInputConnectedChannelsItem];


export const BusinessInputConnectedChannelsItem = {
  WhatsApp: 'WhatsApp',
  Instagram: 'Instagram',
  Email: 'Email',
  Stripe: 'Stripe',
} as const;

export type BusinessInputTheme = typeof BusinessInputTheme[keyof typeof BusinessInputTheme];


export const BusinessInputTheme = {
  green: 'green',
  navy: 'navy',
} as const;

export interface BusinessInput {
  /** @minLength 1 */
  name: string;
  industry: BusinessInputIndustry;
  website?: string;
  location?: string;
  timezone?: string;
  currency?: string;
  description?: string;
  tone?: string;
  avoidKeywords?: string;
  connectedChannels?: BusinessInputConnectedChannelsItem[];
  products?: Product[];
  onboardingComplete?: boolean;
  theme?: BusinessInputTheme;
}

export type BusinessUpdate = BusinessInput;

export interface Error {
  error: string;
}

export interface HealthStatus {
  status: string;
}