# lib/api-zod/src/generated/types/business.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessConnectedChannelsItem } from './businessConnectedChannelsItem';
import type { BusinessIndustry } from './businessIndustry';
import type { BusinessTheme } from './businessTheme';
import type { Product } from './product';

export interface Business {
  id: string;
  name: string;
  industry: BusinessIndustry;
  website: string;
  location: string;
  timezone: string;
  currency: string;
  description: string;
  tone: string;
  avoidKeywords: string;
  connectedChannels: BusinessConnectedChannelsItem[];
  products: Product[];
  onboardingComplete: boolean;
  theme: BusinessTheme;
}