# lib/api-zod/src/generated/types/businessInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessInputConnectedChannelsItem } from './businessInputConnectedChannelsItem';
import type { BusinessInputIndustry } from './businessInputIndustry';
import type { BusinessInputTheme } from './businessInputTheme';
import type { Product } from './product';

export interface BusinessInput {
  /** @minLength 1 */
  name: string;
  industry: BusinessInputIndustry;
  website?: string;
  location?: string;
  timezone?: string;
  currency?: string;
  description?: string;
  tone?: string;
  avoidKeywords?: string;
  connectedChannels?: BusinessInputConnectedChannelsItem[];
  products?: Product[];
  onboardingComplete?: boolean;
  theme?: BusinessInputTheme;
}